// src/infra/enum/channelHeaders-enum.ts
export enum ChennelsHeaders {
    VIDEOS = 'Videos',
    POSTS ='Posts',
    X_POSTS = 'X-posts',
    POSTING_PACE = 'Posting pace / week',
    VIDEO_IMPRESSION = 'Video impression',
    CTR = 'CTR',
    VIEWS = 'Views',
    CPM = 'CPM',
    EARNING = 'Earning',
    AVG_VIDEO_LENGTH = 'Avg video length',
    NET_FOLLOWERS = 'Net followers',
    ENGAGEMENT_RATE = 'Engagement rate',
    VIEWS_VIDEOS = 'Views / video',
}